dikke refactor editting prices want dev had veel changes
