+++ frontend/app/features/archive/components/EventCard.tsx
@@ -1,20 +1,7 @@
 import { useTranslation } from "react-i18next";
-import React, { useState } from "react";
 
-import type {
-  Event,
-  Price,
-  PriceCreate,
-  PriceUpdate,
-} from "~/features/archive/types/eventTypes";
+import type { Event, Price } from "~/features/archive/types/eventTypes";
 import type { Hall } from "~/features/archive/types/hallTypes";
-import {
-  updatePriceByUrl,
-  deletePrice,
-  createPrice,
-} from "~/features/archive/services/eventService";
-import { Protected } from "~/features/auth";
-import { ARCHIVE_PERMISSIONS } from "~/features/archive/archive.constants";
 
 export type EventWithResolvedRelations = Event & {
   resolvedHall: Hall | undefined;
@@ -136,276 +123,12 @@ function EventCardSummary({
   );
 }
 
-type PriceRowProps = {
-  price: Price;
-  eventId: number;
-  onUpdate: (priceUrl: string, updated: Price) => void;
-  onDelete: (priceUrl: string) => void;
-};
-
-function PriceRow({ price, eventId, onUpdate, onDelete }: PriceRowProps) {
-  const { t } = useTranslation();
-  const [isEditing, setIsEditing] = useState(false);
-  const [isSaving, setIsSaving] = useState(false);
-  const [draftAmount, setDraftAmount] = useState<string>(
-    price.amount !== null && price.amount !== undefined ? String(price.amount) : ""
-  );
-  const [draftAvailable, setDraftAvailable] = useState<string>(
-    price.available !== null && price.available !== undefined
-      ? String(price.available)
-      : ""
-  );
-
-  const isModified =
-    draftAmount !==
-      (price.amount !== null && price.amount !== undefined
-        ? String(price.amount)
-        : "") ||
-    draftAvailable !==
-      (price.available !== null && price.available !== undefined
-        ? String(price.available)
-        : "");
-
-  function startEditing() {
-    setDraftAmount(
-      price.amount !== null && price.amount !== undefined ? String(price.amount) : ""
-    );
-    setDraftAvailable(
-      price.available !== null && price.available !== undefined
-        ? String(price.available)
-        : ""
-    );
-    setIsEditing(true);
-  }
-
-  function cancelEditing() {
-    setDraftAmount(
-      price.amount !== null && price.amount !== undefined ? String(price.amount) : ""
-    );
-    setDraftAvailable(
-      price.available !== null && price.available !== undefined
-        ? String(price.available)
-        : ""
-    );
-    setIsEditing(false);
-  }
-
-  async function saveEditing() {
-    setIsSaving(true);
-    try {
-      const updateData: PriceUpdate = {};
-      const parsedAmount = parseFloat(draftAmount);
-      updateData.amount = isNaN(parsedAmount) ? undefined : parsedAmount;
-
-      const parsedAvailable = parseInt(draftAvailable, 10);
-      updateData.available = isNaN(parsedAvailable) ? undefined : parsedAvailable;
-
-      const updated = await updatePriceByUrl(price.id_url, updateData);
-      onUpdate(price.id_url, updated);
-      setIsEditing(false);
-    } catch {
-      window.alert(t("productionPage.price.saveError"));
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  async function handleDelete() {
-    const confirmed = window.confirm(t("productionPage.price.deleteConfirm"));
-    if (!confirmed) return;
-
-    const match = price.id_url.match(/\/prices\/(\d+)/);
-    const priceId = match ? Number(match[1]) : undefined;
-    if (!priceId) return;
-
-    try {
-      await deletePrice(eventId, priceId);
-      onDelete(price.id_url);
-    } catch {
-      window.alert(t("productionPage.price.deleteError"));
-    }
-  }
-
-  if (isEditing) {
-    return (
-      <div className="bg-archive-control flex flex-col gap-2 rounded-lg border border-[color:color-mix(in_srgb,var(--archive-accent)_30%,transparent)] p-3">
-        <div className="flex items-center gap-2">
-          <label className="min-w-[60px] text-[0.6rem] tracking-[0.16em] uppercase opacity-55">
-            {t("productionPage.price.amountLabel")}
-          </label>
-          <input
-            type="number"
-            step="0.01"
-            value={draftAmount}
-            onChange={(e) => setDraftAmount(e.target.value)}
-            placeholder="0.00"
-            className="bg-archive-paper border-archive-ink/10 focus:ring-archive-accent/40 focus:border-archive-accent w-full rounded border px-2 py-1 text-sm focus:ring-2 focus:outline-none"
-          />
-        </div>
-        <div className="flex items-center gap-2">
-          <label className="min-w-[60px] text-[0.6rem] tracking-[0.16em] uppercase opacity-55">
-            {t("productionPage.price.availableLabel")}
-          </label>
-          <input
-            type="number"
-            value={draftAvailable}
-            onChange={(e) => setDraftAvailable(e.target.value)}
-            placeholder="—"
-            className="bg-archive-paper border-archive-ink/10 focus:ring-archive-accent/40 focus:border-archive-accent w-full rounded border px-2 py-1 text-sm focus:ring-2 focus:outline-none"
-          />
-        </div>
-        <div className="flex gap-2 pt-1">
-          <button
-            onClick={cancelEditing}
-            className="text-archive-ink rounded bg-gray-300 px-3 py-1 text-xs font-semibold transition hover:bg-gray-400"
-          >
-            {t("productionPage.edit.cancel")}
-          </button>
-          <button
-            onClick={saveEditing}
-            disabled={!isModified || isSaving}
-            className="bg-archive-accent rounded px-3 py-1 text-xs font-semibold text-white transition hover:opacity-90 disabled:cursor-not-allowed disabled:opacity-40"
-          >
-            {isSaving ? "…" : t("productionPage.edit.save")}
-          </button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="bg-archive-control group/price flex items-center justify-between gap-2 rounded-lg border border-[color:color-mix(in_srgb,var(--archive-accent)_12%,transparent)] p-3">
-      <div>
-        <dt className="text-[0.6rem] tracking-[0.16em] uppercase opacity-55">
-          {t("productionPage.priceLabel")}
-        </dt>
-        <dd className="mt-1 font-medium opacity-95">
-          {price.amount !== null && price.amount !== undefined
-            ? new Intl.NumberFormat(undefined, {
-                style: "currency",
-                currency: "EUR",
-              }).format(price.amount)
-            : t("productionPage.noPrice")}
-          {price.available !== null && price.available !== undefined && (
-            <span className="ml-2 text-xs opacity-60">
-              ({price.available} {t("productionPage.price.available")})
-            </span>
-          )}
-        </dd>
-      </div>
-      <Protected permissions={[ARCHIVE_PERMISSIONS.update]}>
-        <button
-          onClick={startEditing}
-          className="shrink-0 text-[0.62rem] tracking-[0.12em] uppercase opacity-0 transition group-hover/price:opacity-60 hover:!opacity-100"
-          aria-label={t("productionPage.price.edit")}
-        >
-          {t("productionPage.price.edit")}
-        </button>
-      </Protected>
-      <Protected permissions={[ARCHIVE_PERMISSIONS.delete]}>
-        <button
-          onClick={handleDelete}
-          className="shrink-0 text-[0.62rem] tracking-[0.12em] text-red-400 uppercase opacity-0 transition group-hover/price:opacity-60 hover:!opacity-100"
-          aria-label={t("productionPage.price.deleteLabel")}
-        >
-          {t("productionPage.price.deleteLabel")}
-        </button>
-      </Protected>
-    </div>
-  );
-}
-
-type AddPriceFormProps = {
-  eventId: number;
-  onCreated: (price: Price) => void;
-  onCancel: () => void;
-};
-
-function AddPriceForm({ eventId, onCreated, onCancel }: AddPriceFormProps) {
-  const { t } = useTranslation();
-  const [isSaving, setIsSaving] = useState(false);
-  const [draftAmount, setDraftAmount] = useState("");
-  const [draftAvailable, setDraftAvailable] = useState("");
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setIsSaving(true);
-    try {
-      const priceData: PriceCreate = {};
-      const parsedAmount = parseFloat(draftAmount);
-      if (!isNaN(parsedAmount)) priceData.amount = parsedAmount;
-
-      const parsedAvailable = parseInt(draftAvailable, 10);
-      if (!isNaN(parsedAvailable)) priceData.available = parsedAvailable;
-
-      const created = await createPrice(eventId, priceData);
-      onCreated(created);
-    } catch {
-      window.alert(t("productionPage.price.createError"));
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  return (
-    <form
-      onSubmit={handleSubmit}
-      className="bg-archive-control flex flex-col gap-2 rounded-lg border border-dashed border-[color:color-mix(in_srgb,var(--archive-accent)_40%,transparent)] p-3"
-    >
-      <div className="flex items-center gap-2">
-        <label className="min-w-[60px] text-[0.6rem] tracking-[0.16em] uppercase opacity-55">
-          {t("productionPage.price.amountLabel")}
-        </label>
-        <input
-          type="number"
-          step="0.01"
-          value={draftAmount}
-          onChange={(e) => setDraftAmount(e.target.value)}
-          placeholder="0.00"
-          className="bg-archive-paper border-archive-ink/10 focus:ring-archive-accent/40 focus:border-archive-accent w-full rounded border px-2 py-1 text-sm focus:ring-2 focus:outline-none"
-        />
-      </div>
-      <div className="flex items-center gap-2">
-        <label className="min-w-[60px] text-[0.6rem] tracking-[0.16em] uppercase opacity-55">
-          {t("productionPage.price.availableLabel")}
-        </label>
-        <input
-          type="number"
-          value={draftAvailable}
-          onChange={(e) => setDraftAvailable(e.target.value)}
-          placeholder="—"
-          className="bg-archive-paper border-archive-ink/10 focus:ring-archive-accent/40 focus:border-archive-accent w-full rounded border px-2 py-1 text-sm focus:ring-2 focus:outline-none"
-        />
-      </div>
-      <div className="flex gap-2 pt-1">
-        <button
-          type="button"
-          onClick={onCancel}
-          className="text-archive-ink rounded bg-gray-300 px-3 py-1 text-xs font-semibold transition hover:bg-gray-400"
-        >
-          {t("productionPage.edit.cancel")}
-        </button>
-        <button
-          type="submit"
-          disabled={isSaving}
-          className="bg-archive-accent rounded px-3 py-1 text-xs font-semibold text-white transition hover:opacity-90 disabled:cursor-not-allowed disabled:opacity-40"
-        >
-          {isSaving ? "…" : t("productionPage.price.addLabel")}
-        </button>
-      </div>
-    </form>
-  );
-}
-
 type EventCardProps = {
   event: EventWithResolvedRelations;
-  onPricesChanged?: (eventId: string, prices: Price[]) => void;
 };
 
-export function EventCard({ event, onPricesChanged }: EventCardProps) {
+export function EventCard({ event }: EventCardProps) {
   const { t, i18n } = useTranslation();
-  const [prices, setPrices] = useState<Price[]>(event.resolvedPrices);
-  const [isAddingPrice, setIsAddingPrice] = useState(false);
 
   const getLocalizedHallName = (hall?: Hall): string => {
     const lang = i18n.language.startsWith("nl") ? "nl" : "en";
@@ -420,31 +143,12 @@ export function EventCard({ event, onPricesChanged }: EventCardProps) {
   const dateAndTime = formatEventDateTimeRange(event.starts_at, event.ends_at);
   const eventDate = dateAndTime?.dateLabel ?? t("productionPage.fallback.dateUnknown");
   const eventTime = dateAndTime?.timeLabel ?? t("productionPage.fallback.dateUnknown");
-const eventLocation = getLocalizedHallName(event.resolvedHall ?? event.hall);
-
-  const match = event.id_url.match(/\/events\/(\d+)/);
-  const eventId = match ? Number(match[1]) : undefined;
-
-  function handlePriceUpdate(_priceUrl: string, updated: Price) {
-    setPrices((prev) => prev.map((p) => (p.id_url === updated.id_url ? updated : p)));
-    const updatedList = prices.map((p) => (p.id_url === updated.id_url ? updated : p));
-    onPricesChanged?.(event.id_url, updatedList);
-  }
-
-  function handlePriceDelete(priceUrl: string) {
-    const updatedList = prices.filter((p) => p.id_url !== priceUrl);
-    setPrices(updatedList);
-    onPricesChanged?.(event.id_url, updatedList);
-  }
-
-  function handlePriceCreated(created: Price) {
-    const updatedList = [...prices, created];
-    setPrices(updatedList);
-    setIsAddingPrice(false);
-    onPricesChanged?.(event.id_url, updatedList);
-  }
-
-  const hasPrices = prices.length > 0;
+  const eventLocation = getLocalizedHallName(event.resolvedHall ?? event.hall);
+  const eventPrice = formatPrices(
+    event.resolvedPrices,
+    i18n.language,
+    t("productionPage.noPrice")
+  );
 
   return (
     <li>
@@ -459,61 +163,10 @@ const eventLocation = getLocalizedHallName(event.resolvedHall ?? event.hall);
         <div className="border-t border-[color:color-mix(in_srgb,var(--archive-accent)_14%,transparent)] px-4 py-3 md:px-5">
           <div className="grid items-start gap-3 text-sm sm:grid-cols-2">
             <EventCardDetail label={t("productionPage.timeLabel")} value={eventTime} />
-{hasPrices ? (
-              <div className="flex flex-col gap-2">
-                {prices.map((price) => (
-                  <PriceRow
-                    key={price.id_url}
-                    price={price}
-                    eventId={eventId!}
-                    onUpdate={handlePriceUpdate}
-                    onDelete={handlePriceDelete}
-                  />
-                ))}
-                {isAddingPrice && (
-                  <AddPriceForm
-                    eventId={eventId!}
-                    onCreated={handlePriceCreated}
-                    onCancel={() => setIsAddingPrice(false)}
-                  />
-                )}
-                <Protected permissions={[ARCHIVE_PERMISSIONS.create]}>
-                  {!isAddingPrice && (
-                    <button
-                      onClick={() => setIsAddingPrice(true)}
-                      className="text-archive-accent text-[0.62rem] tracking-[0.14em] uppercase opacity-70 transition hover:opacity-100"
-                    >
-                      + {t("productionPage.price.addLabel")}
-                    </button>
-                  )}
-                </Protected>
-              </div>
-            ) : (
-              <div className="flex flex-col gap-2">
-                {isAddingPrice ? (
-                  <AddPriceForm
-                    eventId={eventId!}
-                    onCreated={handlePriceCreated}
-                    onCancel={() => setIsAddingPrice(false)}
-                  />
-                ) : (
-                  <div className="flex items-center gap-2">
-                    <EventCardDetail
-                      label={t("productionPage.priceLabel")}
-                      value={t("productionPage.noPrice")}
-                    />
-                    <Protected permissions={[ARCHIVE_PERMISSIONS.create]}>
-                      <button
-                        onClick={() => setIsAddingPrice(true)}
-                        className="text-archive-accent shrink-0 text-[0.62rem] tracking-[0.14em] uppercase opacity-70 transition hover:opacity-100"
-                      >
-                        + {t("productionPage.price.addLabel")}
-                      </button>
-                    </Protected>
-                  </div>
-                )}
-              </div>
-            )}
+            <EventCardDetail
+              label={t("productionPage.priceLabel")}
+              value={eventPrice}
+            />
             <EventCardDetail
               label={t("productionPage.address")}
               value={
