+++ frontend/app/locales/en/translation.json
@@ -176,7 +176,7 @@
       "delete": "Delete info (en)",
       "confirm": "Are you sure you want to delete this production info? This action cannot be undone.",
       "error": "Error: Each production should have info in at least one language."
-},
+    },
     "add": {
       "add": "Add production info"
     },
@@ -191,7 +191,6 @@
       "amountLabel": "Amount",
       "availableLabel": "Available",
       "available": "available"
-    }
     }
   },
   "home": {
