+++ frontend/app/locales/nl/translation.json
@@ -176,7 +176,7 @@
       "delete": "Verwijder info (nl)",
       "confirm": "Weet je zeker dat je deze productie-informatie wilt verwijderen? Deze actie kan niet ongedaan worden gemaakt.",
       "error": "Error: Elke productie heeft info in minstens één taal."
-},
+    },
     "add": {
       "add": "Voeg productie-informatie toe"
     },
@@ -191,7 +191,6 @@
       "amountLabel": "Bedrag",
       "availableLabel": "Beschikbaar",
       "available": "beschikbaar"
-    }
     }
   },
   "home": {
